requirements.txt
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==2.0.0
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
django-model-utils==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
gpg===1.13.1-unknown
gunicorn==20.1.0
httplib2==0.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
monotonic==1.5
mysql-connector-python-rf==2.2.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
paramiko==2.6.0
pbr==5.11.1
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.1.0
protobuf==3.6.1
psycopg2==2.9.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
redis==4.5.5
reportlab==3.5.34
requests==2.30.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.19.1
six==1.14.0
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
typing_extensions==4.6.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.2
usb-creator==0.3.7
virtualenv==20.20.0
wadllib==1.3.3
xkit==0.0.0
